Unit & e2e tests

+++ CashFlow/src/CashFlow.Application/UseCase/Expenses/Delete/DeleteExpenseUseCase.cs
@@ -1,5 +1,6 @@
 ﻿using CashFlow.Domain.Repositories;
 using CashFlow.Domain.Repositories.Expenses;
+using CashFlow.Domain.Services.LoggedUser;
 using CashFlow.Exception;
 using CashFlow.Exception.ExceptionsBase;
 
@@ -7,25 +8,37 @@ namespace CashFlow.Application.UseCase.Expenses.Delete;
 
 public class DeleteExpenseUseCase: IDeleteExpenseUseCase
 {
+    private readonly IExpensesReadOnlyRepository _expenseReadOnly;
     private readonly IExpensesWriteOnlyRepository _repository;
     private readonly IUnitOfWork _unitOfWork;
+    private readonly ILoggedUser _loggedUser;
 
     public DeleteExpenseUseCase(
+        IExpensesReadOnlyRepository expensesReadOnlyRepository,
         IExpensesWriteOnlyRepository repository,
-        IUnitOfWork unitOfWork)
+        IUnitOfWork unitOfWork,
+        ILoggedUser loggedUser)
     {
+        _expenseReadOnly = expensesReadOnlyRepository;
         _repository = repository;
         _unitOfWork = unitOfWork;
+        _loggedUser = loggedUser;
     }
 
     public async Task Execute(long id)
     {
-        var result = await _repository.Delete(id);
+        var loggedUser = await _loggedUser.Get();
 
-        if (result == false)
+        var expense = await _expenseReadOnly.GetById(loggedUser, id);
+
+        if (expense is null)
         {
-            throw new NotFoundException(ResourceErrorMessages.EXPENSE_NOT_FOUND);
+            throw new NotFoundException(ResourceErrorMessages
+                .EXPENSE_NOT_FOUND);
         }
+
+        await _repository.Delete(id);
+
 
         await _unitOfWork.Commit();
     }
